Begin setting up NuGet package
Adding additional details and tags, and connecting and filling in the things I need to fill! For some reason, NuGet was upset about where the Microsoft.Windows.Shell dll ends up located at; I simply tried running the project without it and it seemed to work just fine! Time will tell if there's a reason that I need to come back to fix that.