Enabled Source Link
Figured out what I needed to do to enable Source Link for my package. Note that I didn't follow the recommended procedure for ensuring a deterministic build, due to the fact that I don't use a CI system. I never debug using the release version anyway, so hopefully it won't be an issue.