Major improvements to TabControl
The Tab Control is now practically done in terms of core functionality. There are a few things still missing, such as updating the tabs themselves if the TabItem data is changed, but that will come soon. Additional functionality (such as an "Add Tab" button and a tab list menu a la Visual Studio are also planned before this control is fully finished (for its first iteration). Tab scrolling works almost perfectly when the tab list is wider than can be shown.

Added a TabControlTest page to the SsuiSample application, which I can use to test out the TabControl.