FontSelectDialog methods to load from/set to control properties
Added LoadFromControl and SetToControl methods, to load in selected font values from a control, or to set a control with selected font values. Has special handling for TextBox and TextBlock for TextDecorations, as that property is not common to all controls. Note that no handling is done based upon what is visible in the UI of the dialog itself, these methods simply attempt to load and set all properties it can. If this is a problem, these methods should not be used.