Changed Menu to DependencyProperty
Currently dealing with issues with the MenuButton and the context menu its in not binding to an ItemsSource, so I'm trying to figure out what's going on. Thought this would help. It is not.

Either way, this may be a win for anyone who needs this. Functionality of MenuButton is otherwise unchanged.