Created templated TabControl, almost ready!
The templated TabControl has about 90% of the functionality of the control currently in the Experimental namespace (I need to add the functionality for the scroll buttons). It took a lot of trial and error, but it finally works! I've been messing with this on and off the past few days or so. Once the remaining functionality is implemented and some visual bugs fixed, this new TabControl will be fully ready to go!

Updated the TabControlTest to use the new TabControl. I had to duplicate TabDisplayItem while I was fixing some bugs, but I was able to reuse TabItem from the Experimental version.

Some properties and such are "internal protected". I didn't really want to have code with such a access level, but I couldn't think of  a better way to connect the TabDisplayItem to the TabControl. I don't like the idea of "special code" that can't be modified or accessed in case someone wants to inherit my controls or classes or something. At least with the "protected" part, that helps alleviate my concerns somewhat. By the way, it is very hard to actually access the visual elements (the TabDisplayItems) from an ItemsControl, that's very annoying.