Add tab bar scrolling functionality
The tab bar will now scroll when the scroll buttons are clicked. This is achieved via a new TabBarScrollCommand command. While this is meant to be called internally, it can also be used by other apps if so desired.

Also, the tab bar will now select the first tab automatically when loaded. This functionality can be disabled, if desired, via the SelectFirstTabOnLoad property. Note that the property will only make a difference when used in XAML or prior to the control loading; changing the property afterwards will do nothing.