Add ShowIcon property, padding on right
Add padding to the right side of the tab, to give a bit of space between the tab title and the close button of the tab. Added a ShowIcon property; if false, the icon area of the tab is hidden (even if there is an icon set). With ShowIcon and CanClose set to false, the tabs will look more like they do in the WPF TabControl (when only a label or TextBlock is used).