Add more events and methods for closing
Added Close command to the TabItem, which will tell the TabControl to attempt closing it. Closing a tab via any method will now raise a TabClosing event, which can allow a developer to cancel closing it. The TabControl also can perform a set action when you close the selected tab: either not selecting anything, or selecting tabs to the left or right. Added the ability to test this via the test window.