Fix bugs with DoubleSpinner and IntegerSpinner timer
Added an error handler in the timer to make sure that some weird messup can't cause that to crash the app. Also fixed a bug where if the mouse left the control while holding down the up or down button, it would continue going up or down endlessly. This means that the mouse leaving the control at all will cancel it, which may not be the best behaviour, but for now, it prevents the buggy behaviour. In a future release, I could tweak this further.