Tab drag-and-drop now works!
Dragging and dropping tabs works now! Dragging a tab over another tab will make some placement indicators appear, showing where the tab will go when you drop it. There is work that I still need to do as it doesn't fully work as expected, especially with only 2 tabs in the list. In the future, I will add an option to disable drag-and-drop as desired.