Disable cross-TabControl drag-and-drop
In the future, I may allow a way to have drag-and-drop across TabControls, but for right here and now, it will be disabled.

The biggest issue right now is that when a tab is dragged over, the original TabControl has no way of knowing it can remove the tab; as a result, the tab and its contents are duplicated (or worse, an exception occurs as the TabItem's content already has a parent). I think in the future this can be allowed by creating a custom IDataObject class to use in drag-and-drop, and in the custom class I can mark what the original TabControl is.

Right now, I don't want to open that can of worms, and I'd rather leave it for a future version where I can expand and improve on this further.