TabItem: Drag events, BringIntoView, ToolTip
Added a number of features to TabItem, including the ToolTip property, events for all of the drag and drop operations (DragEnter, DragOver, DragLeave, Drop, and previews), added BringIntoView method. Some drag events may raise when dragging a tab, which shouldn't be the case. I'll need to fix that before the final release.

Added ToolTip to the first tab in the test window, and also added a small display to show off drag events being raised on the first tab.