Added visual element for TabItem.IsDirty
Added property DisplayDirtyState to TabItem, to set if the TabItem should display the state of its IsDirty property in the UI. By default, this is true; when a tab is dirty, the asterisk (*) is added to the tab title in the tab bar and the tab list menu.

Added the ability to toggle the IsDirty state for any given tab in the test window.