Refactor/improve drag-and-drop
Fixed issues where drag events would raise while dragging tabs (which shouldn't be the case). Also fixed issues where some drag events would not raise when AllowTabDragDrop was set to false but a tab's AllowDrop was true. On top of that, brought consistency to the code for all of the drag-and-drop handling code here. All in all, this should be good!