Fix visual issue when drag-and-dropping tabs
Sometimes when dragging a tab around, it will no longer have the IsSelected state set correctly. This is now fixed (by going through all tabs after a drag-and-drop and setting their states lol). There may be a slightly faster way to do this, but for here and now, this works!