Further work with FlatWindow RoundedCorners
The more I'm researching into this, the rounded corners feature may not be possible with FlatWindow, without a lot of extra work that I personally don't want to put the time into doing.

What I've added here is some extra native methods that might be able to allow this to happen. However, if this proves to not work or if it causes weird side effects or further issues, then I won't continue with this endeavor.

Reference: https://web.archive.org/web/20190112214258/https://blogs.msdn.microsoft.com/wpfsdk/2008/09/08/custom-window-chrome-in-wpf/