Added LeftTabBarElement and RightTabBarElement
Similar to the TopLeftElement and TopRightElement properties in FlatWindow, these two new properties (LeftTabBarElement and RightTabBarElement) allow you to place UI elements on the left and right side of the tab bar in a TabControl. This can be used to, say, create your own "Add Tab" button or put a menu in the top-left of the tab bar.

Also rearranged and cleaned up the TabControl file a bit as well.