Disable CornerRadius in FlatWindow
I still haven't figured out how to get corner radius to work exactly with FlatWindow. However, I was given another option that I can mess with later. I'm not going to try to get this out for this current 1.8.3 release, and I'll instead save it for 1.9 if I can get it to work.