Fix icons not appearing with MessageDialog
I felt like I fixed this issue like 2 times before, but apparently not! This came down to a problem of the old name of this library being used in place of the current one. Kind of embarrassing I let this issue go for so long.

If this somehow gets broken again, I'm going to assume my code is cursed.