Removed unnecessary usings
Removed unnecessary using statements in IntegerSpinner and DoubleSpinner. Also updated warning level settings for the entire project.

I've decided to cancel the idea to rebase IntegerSpinner and DoubleSpinner under the RangeBase class; the benefits I would get would be minimal, and instead it would add complexity, especially for the IntegerSpinner.