Create NewSelectPanel that is ControlTemplate-based
Once the NewSelectPanel control completely replicates all features of the original SelectPanel, this new control is fully replacing the old one. The NewSelectPanel is ControlTemplate-based, rather than being a ContentControl. This does a number of things, including - most importantly - allowing subitems to be added in the designer or XAML. The new control is also backed by the SelectableCollection class, which helps bring more type safety.

This first commit contains most of the functionality, as well as a lot of the visuals, but it still has a number of bugs and isn't quite yet a drop-in replacement for the original SelectPanel (due to missing a lot of convenience methods, such as "SelectAll"). Future commits will focus on first fixing the biggest bugs and then adding in the remaining functionality and methods.