Somewhat fix selection issues in NewSelectPanel
Basic selection should now work properly, including selecting with the Ctrl key if MultiSelect is true. The SelectAll method doesn't seem to work unless you're holding down the Ctrl key, so that's another bug that will need to be fixed.

Also, the internal count of selected items isn't entirely accurate.

Right now, the SelectableCollection's SelectedItems list isn't connected in any way to the actual items that are being selected in the UI, so right now I programmatically update the SelectedItems list whenever something is selected or deselected, but this feels very clumsy and is clearly already causing problems. I'll need to rethink this.