Add most remaining functions from SelectPanel
Most of the remaining convenience functions from SelectPanel, as well as the functions for moving items in the list were brought over. Some functions will not make the transition, such as RemoveRange, due to there not being a convenient equivalent function in ObservableCollection.

In the original SelectPanel, people were discouraged from directly using the Children property due to the custom code that was handling the items, leading to these convenience methods being created. That is no longer the case with this NewSelectPanel, so these methods now solely exist for convenience and backwards-compatibility. Part of me wants to remove many of these methods, actually, but for now, they will stay. If I change my mind, I may mark them as Obsolete.

There is currently a change in behavior with the MoveSelectedItemsUp and MoveSelectedItemsDown functions, where the selection is cleared after the move, so the items need to be re-selected if you want to move them again. This will be fixed before the release.

At this point, nearly all functionality has been ported over from the original SelectPanel. Expect it to be replaced in one of the next few commits.