Create PropertyList control (experimental)
Created the PropertyList control. This control will be in its own namespace, called SolidShineUi.PropertyList. There will be a lot of classes and controls, and so I don't want the main namespace cluttered with all of that.

Right now, the control is actually called ExperimentalPropertyList, as it is still, well, experimental. As of right now, it can load an object and list all of its properties, and get the ToString result of the properties.

In the future, I will expand upon this to allow the PropertyList to properly display more values, and then allow them to be edited. This will take time though.