PropertyList add View menu, fix category sort
In the PropertyList control, replaced the two sort buttons with a "Sort and View" menu. I will put an icon on it later. Added an option in the menu for showing and hiding the Type column; the column is now hidden by default.

Made improvements to the category sort function, so that properties without a category are placed at the bottom of the list, and properties in the same category are sorted alphabetically.