Add PropertyEditors for PropertyList
The PropertyEditors (and associated IPropertyEditor interface) is what will display the various property values in a way that can be viewed, understood, and edited. For now, I've added editors for a handful of basic types, but a lot more will be on the way in time.

I'm currently having the issue where the actual editors are waaaaaaaaay larger than the area they're supposed to be fitting into, so I'm going to need to investigate this and fix it.