PropertyList can edit properties, add EnumEditor
Added extra members to IPropertyEditor to allow editing. When you change a value in any of the existing editors, it'll automatically update on the target object. I also added a little bit of error handling for when attempting to edit a property that can't be edited. In the future, I'll add some visual feedback to show that a property is read-only.

Added in the EnumEditor, which PropertyList does a little bit of extra handling to hook up and use properly. It will display any Enum; if, for whatever reason, someone wants to replace the EnumEditor with their own EnumEditor, they can do so by registering an editor for the Enum type.