DoubleEditor supports Half type, add BrushEditor
Added support for the System.Half number type in the DoubleEditor. This type is only available in .NET 5 and 6. Thank goodness for preprocessor if statements!

Added a first draft of the BrushEditor. It currently doesn't actually allow editing, but it's a start. I'm probably going to redesign the UI, as I'm not happy with the direction I'm going here.