Add LightTheme and DarkTheme color schemes
Having missed the trend up until now, I've added CreateLightTheme and CreateDarkTheme static methods to ColorScheme. These create a white-based and black-based color scheme, similar to the Light mode and Dark mode present in Windows 10/11 and various other programs and sites. An accent color can be passed in to add a bit of color at certain points; if you want to make certain UI elements stand out all the time while using these themes, pass in the accent color and also use the "UseAccentColors" property on that UI element.

I also disabled the PanelsUseAccent property in ColorScheme, as that was never used. MenusUseAccent is also never used in ColorScheme, but I plan to address this.

Updated UI tests to fix their visual issues. Added the ability to set the color scheme of the test window to a high contrast theme or the new light or dark themes.