Redesign BrushEditor, add FontFamilyEditor
Began a redesign for the BrushEditor for PropertyList. Now there is a row of 6 buttons, and the relevant button is highlighted when that type of brush is set. Still a decent amount of work to do, but this is the direction I've decided I wanted to go with its UI.

Added in the FontFamilyEditor, which is for... editing the FontFamily type. Uses the FontSelectDialog to set a font. This editor technically doesn't handle null values as much as the other editors, but for now, I think it's okay. I'm sure there's probably some situation where setting a FontFamily property to null does something, but I'll deal with that when that time comes.