Add FontStyleEditor for PropertyList
Another editor... this time for the FontStyle class. WPF's FontStyle class is kinda weird, because it only realistically has three values: Normal, Italic, or Oblique, but yet this is not represented as an Enum, but instead has static objects FontStyles.Normal, FontStyles.Italic, and FontStyles.Oblique.

Unfortunately, this means that I can't really ever guarantee that I cover all possible values for FontStyle. Thus, I set FontStyles to Normal if it's a value that isn't one of the three I listed above.