Add UriEditor, GuidEditor
Some of these more specialized types aren't actually too hard to develop editors for, as I copy the existing StringEditor (or FontFamilyEditor) and then tweak them to fit the new type.

UriEditor is for storing and displaying Uris, as you would expect. It includes the ability to load a Uri by selecting a file. Note that some Uris are WPF resource Uris, which may mess with the "Open URL" function. If you try typing a Uri into the textbox that isn't valid, your text is wiped and replaced with the last valid Uri value when you leave the box.

GuidEditor doesn't allow directly editing the Guid in the editor, similar to the FontFamilyEditor. I do give users a way to directly set the Guid via the "Set Guid" command, but the New button is most prominent right in the UI, as I feel that will be more useful for people. Accidentally hitting the button and generating a new Guid may be destructive though, given that Guids may be used for important purposes, so I might move the "New" button into the menu as well, just so it then requires two clicks (since hitting "New" can't be reversed).

Fixed a bug in the FontFamilyEditor for .NET Framework (where the button wouldn't work).