Add BrowseButtonText property to FileSelect
I originally wanted this to be a DependencyProperty, but for some reason this just wasn't working. I decided to just make this a basic property. This property allows the text in the Browse button to change to something other than "Browse...", i.e. for the purposes of localization or what-not. Prior to the PropertyList, this was the first control that had text that couldn't be directly edited.