Create RelativePositionSelect control
Created a new control in the Utils namespace. It is very useful only in niche situations, which is why it isn't in the main namespace.

This control presents a box with a circle in it, and you're able to select a double value between 0.0 and 1.0 in both the X and Y axis by moving the circle. Snap lines can also be added, which the circle will automatically snap to once within a certain distance.

This was designed for use in setting the StartPoint and EndPoint properties in a LinearGradientBrush, and the GradientOrigin property of a RadialGradientBrush. You'll see these use cases once I build the gradient editor dialogs.

Also added a test panel for the sample app, although this will be removed later.

I'm proud of my work with this, but there's still a few bugs to fix. For the most part, it is ready though!