Fix visuals/bugs with RelativePositionSelect snap lines
Fixed the bug with the selector not lining up correctly with the snap lines. Swapped the colors of the snap lines and selector, when applying a color scheme, to make the selector stand out more and the snap lines stand out less.

Fixed other bugs in relation to the snap lines not updating in color when the SnapLineBrush property is set directly, and the selector not updating when the control is resized. Also moved the snap lines to render behind the selector.

At this point, this control is now fully finished and ready.