Add keyboard support to RelativePositionSelect
This is the final feature I'll be adding to this control.

Added support for the selector to be moved via the keyboard. By pressing the up, down, left, and right arrow keys, the selector is moved while the control is in focus. You can also use the Home, End, Page Up, and Page Down keys to move the selector to the far edges of the control as well.

The distance the selector moves via the arrow keys is set via the KeyMoveStep property (maybe I'll rename this later?), and the color of the keyboard focus highlight (the outline that appears when the control has focus) is set via the KeyboardFocusHighlight property.

The PreviewKeyDown event is used to prevent WPF using the arrow keys to try to move the focus to other controls.