Changed DoubleEditor Step value to 0.5, instead of 1
Maybe I could add in a menu option to it to allow end users to change the Step value. Primarily, my concern here was that the Opacity property of UIElements is a double, and a step of 1, clicking up or down just goes from full transparency to none; setting the step to 0.5 at least puts in a middle point.