Add ThicknessEditor, arrow icons
Add a ThicknessEditor for the PropertyEditor control. Added in all kinds of arrow icons to use for this editor and some others in the future.

This editor is actually pretty wide (along with the BrushEditor), and so I expanded the width of the PropertyList control in the test program. I think the minimum suggested width will be 400 pixels from now on, which is a lot. I tried to make this editor a bit more flexible in width by allowing the extra space by the arrow icons to collapse, but I didn't want the arrows to slide under the spinners, so I could only do so much.