More features for TabItem and TabControl
Added more properties to TabItem: TabBackground (brush for the background of the tab iteself) and Padding (space between the title/icon and the rest of the tab).

Also fixed the issue that CanSelect property didn't actually do anything lol.

Added the HorizontalTabBarHeight property to TabControl, to allow the height of the tab bar itself to be changed (rather than being set to 24). Note that heights under 20 may result in a slightly undesirable appearance. Adjusted the templates for TabControl and TabDisplayItem to support the height change.

Also set the ContentProperty of TabControl to the Items property, so you can directly define the child TabItems in XAML.