Changed Items property to be read-only
Changed the Items property in TabControl and NewSelectPanel to be read-only, with the Items not being able to be set to a new collection. This avoids issues where one could set a new collection, and then cause issues with how the collection interacts with the parent control.

Also updated ColorPickerDialog to work with this new change. Thus, this change is a breaking change from 1.8.4.