Change tab close image to be XAML path (glyph)
Change the Tab close drawing to be an actual XAML path (or glyph), rather than being an image. This allows the glyph to be colored correctly in high contrast color schemes, and also avoids any issues with things looking pixel-y at higher DPIs or zoom sizes.

In the future, I want to offer a way to allow users to change the color of the close glyph independently of setting the ColorScheme, but at least this is better than before.