Rewrite FileSelect control, add supporting classes
The FileSelect control was rewritten to be a Control with a template and styling, rather than a UserControl like before. This will allow developers to create their own custom styles for the control as well. The control now performs more predictably (and with less caveats) and there's some nice new features too.

The "Browse" button can be placed on any side of the control, or even be hidden. The "(no files selected)" message can be customized. Multiple files can be displayed as a list, rather than an "x files selected" message appearing.

A few properties were removed, such as "AllowContextMenu" and "PressOnClick". Some were just irrelevant, did nothing, or were redundant, while others like "AllowContextMenu" can now be dealt with in more conventional means, by changing the ContextMenu property of the control itself.

With this new control, some helper classes were also added:

1. FilenameStringCollection, which is used for the SelectedFiles property. This is an ObservableCollection, with the ability to limit the maximum collection size and an AddingItem event to allow cancelling the addition of a particular item.
2. FilePathToIconConverter, a XAML converter that will take a file path and return out the file icon in a 16x16 image.
3. FilePathToNameConverter, a XAML converter that will return just the filename of a full file path.
4. ImageTextListItem, which can display some text and an image, along with a remove button. This can also be used in a SelectPanel.

I also updated the NewSelectPanelTest to utilize the ImageTextListItem.

In a future commit, I will remove the existing FileSelect control and rename "NewFileSelect" to just "FileSelect".