Disallow duplicates in NewFileSelect, fix control sizing
A file cannot be added into the FilenameStringCollection now if it was already present in the collection.

Changed the minimum height of the Browse button and the margin of the "no files selected" text in the NewFileSelect. In the future, I'll probably add some properties to more control the size and positioning of the button.