Renamed FilenameStringCollection to LimitableStringCollection
There isn't anything about the collection that makes it specific to filenames, per se, so I'd rather select a more generic name. Right now, I'm only using it in FileSelect, but it can be useable in other situations too.

Also renamed the Capacity property to MaxCount, and added a little bit of XML comments (more to come!). Also set the Content property of FileSelect to be the SelectedFiles collection; in theory, this should allow developers/designers to preset selected files if this is desirable.