Deprecated old SelectPanel helper methods
With access to the Items and Items.SelectedItems methods now available and preferred, these helper methods are now marked as obsolete. They won't be removed for a little while, but I'd rather mark them obsolete now to prepare consumers of this library.

Also added more XML comments to properties and methods missing them.