Adds PropertyList to FlatWindowTest
Added a PropertyList to the FlatWindowTest, so that testers can modify various properties of the opened FlatWindow. The PropertyList is openable once a FlatWindow is created, and automatically closes once the FlatWindow is closed.

With this, I added the requirement that only one FlatWindow can be opened at a time through FlatWindowTest, and so if you attempt to open a second one, it simply focuses the first one. Also added code that closes the FlatWindow if the main test app is closed.