Added ItemAdding event to SelectableCollection
I debated back and forth about adding this in, but I decided, why not lol. This event will allow implementers to inspect an item before it gets added to the collection, and potentially cancel adding the item if desired.

One example may be to check to make sure the item being added doesn't already exist in this collection; if it does, then don't add it again.

This is similar to the AddingItem event in the LimitableStringCollection (which I renamed to ItemAdding for consistency). The difference is that the event in the SelectableCollection now supports generics, rather than just a string.