Fix move up function in NewSelectPanel
I may have went overboard with my solution, but it seems to work consistently so that's what matters for right now.

Fixed the move up function in NewSelectPanel, so that the relative index of multiple items being moved does not change. This fix was done by sorting all of the moved items by their index in the parent collection; this may be a bit overkill, but the end result is consistent.

Also added a shortcut if only 1 item is selected, so that we don't need to deal with sorting and all this stuff in that case.

Also changed LoadTemplateItems to prepare for the next commit (oops lol).