Reimplement AllowParentScrolling in NewSelectPanel
Reimplement the AllowParentScrolling property from the original SelectPanel control. In order to implement it here, I needed to connect to the ScrollViewer from the template. Thus, this property will not work if a developer writes a new style that doesn't name the controls.

I know I had situations in the past where I utilized this property, but in my current projects, I couldn't find any instances of it. So I'm not sure how useful this is to me or others, but I brought it over for the sake of compatibility (and also there's no harm in having it implemented here at all).