Fix KeepMenuArrowOnRight property in MenuButton
When this property was set to True before, it also overrode the HorizontalContentAlignment to always be Center, which wasn't ideal. I redesigned how the arrow is drawn, and so now this issue won't appear!

Also fixed the issue where the arrow will not update its color in high contrast themes, or if the control is disabled.