Add support for PowerToys exported color lists
Microsoft PowerToys 0.57 added support to export the list of colors in its Color Picker tool to a file. Files can be exported in a plain text ".txt" file or a JSON ".json" file, and the data sorted by each color or by the selected color formats.

This code only supports the ".txt" files for right now, but supports both sorting methods. Ideally, as long as the format doesn't get expanded or messed with in future versions of PowerToys, this will continue to work.

Also, didn't realize this, but local functions was actually added into C# version 7.0, which .NET Framework can support!