Fix occasional visual glitch in FileSelect
When set up in the right way, the background of the FileSelect control would look weird while the mouse interacted with it. This is because the highlight and click brushes of the internal ImageTextListItem controls were set to match the background of the entire control, which led to the weird results on mouse-over.

Now this has been corrected by using an old favorite of mine: a barely visible white brush!