Redesign Styles/Decorations area of FontSelectDialog
Replace the standard checkboxes in the FontSelectDialog with SSUI ones, so that the text is colored appropriately under all color schemes.

After changing the layout of the controls, this also ultimately led to this section being a little bit shorter than it was before! More room for showing the big font list!