Added XAML commands to TabControl
Added some XAML commands that developers can use with the TabControl. The commands are "CloseCurrentTab", "CloseSpecificTab", and "SwitchToTab". The function of these commands should make sense given their titles.

TabControl already utilizes a few commands for some functions internally, but these commands were built with the intention that anyone can use them. In fact, I removed one of the existing commands for the tab list menu, and am now using SwitchToTab for that instead.

Adding in the CloseSpecificTab command also opens the possibility for me to add a close option to the menu if I customize the menu items' appearances even more. Which is possible.