Add ItemsSource property to SelectPanel
With ItemsSource, you can supply your own ObservableCollection or SelectableCollection of controls to use with the SelectPanel instead. Rather than exposing the property as a generic IEnumerable, I specified it needs to be a IEnumerable<SelectableUserControl>, to match the type needed for the SelectPanel to work correctly. Perhaps in the future I can work out a way to allow a less restrictive data type here.

With this being set, now the SelectPanel intializes with its own Items property being set as the ItemsSource by default.

With the ItemsSource, you should then be able to use the CollectionViewSource property to add WPF sorting and filtering, if desired.