Add ISelectableCollection, breaking change
Added an interface ISelectableCollection that SelectableCollection implements, which I can also use for something else in a future commit.

Unfortunately, due to fun stuff with using classes with generic types, I had to change the event handler for SelectableCollection's SelectionChanged event. This makes a breaking change. I'll have to disclose this with the next release.

Expect the next release to simplify and rearrange some stuff with SelectPanel.