Create SelectableCollectionView
This is a new class that inherits from CollectionView, and also implements the new ISelectableCollection.

I actually renamed ISelectableCollection to ISelectableCollectionView, and removed the inheritance on ICollection and INotifyCollectionChanged from the interface. That was because the CollectionView class doesn't offer a way for me to implement anything from those base interfaces.

I've tried to do the best I can to allow as many types as possible in the SelectableCollectionView, despite the generic type present as well. Trying to use the SelectableCollectionView with an IEnumerable that doesn't also notify on collection changes (INotifyCollectionChange) may lead to some wonky situations.