Rename SelectableCollectionView to SelectableCollectionViewSource
Semantics! ... To be honest, I don't fully understand the difference between a CollectionView and a CollectionViewSource, beyond the simple fact that a CollectionViewSource can be used in XAML and a CollectionView, seemingly, cannot be? *shrug*

Given that my goal is to use this in XAML and WPF, I've opted to base this class off the CollectionViewSource instead.