Create ListEditor, EnumerableEditor for PropertyList
These editors are right now pretty barebones, but the idea is that it replaces the kind of ugly "System.Collections.Generic.List`1" text. The ListEditor will display the number of items in the list; the EnumerableEditor will not.

In a future commit, I'll create a window to view (and maybe edit) the items in said lists and enumerables.

Support for Dictionary and other generic types are not present.