Add various appearance properties to PropertyList
Added the dependency properties ObjectDisplayName, ShowReloadButton, ShowViewMenu, and ShowFilterBox.

The ObjectDisplayName property replaces the previous code for displaying the names of observed objects, and allows developers to set a custom name instead.

The various Show properties are for displaying UI elements in the toolbar at the top of the control. Hiding all elements will hide the toolbar entirely.

Also made some layout adjustments.