Add comments to TabDisplayItem, breaking change
Added missing XML comments to TabDisplayItem. Technically, this is mostly an internal object, so comments may not be necessary, but I'd rather just get it done now anyway.

Removed the ApplyColorScheme(HighContrastOption) method overload. Please just use the standard ApplyColorScheme method instead, with ColorScheme.GetHighContrastTheme. This same change will be applied to the remaining controls in the future.