Add GiveFeedback, QueryContinueDrag events to TabItem
These events don't seem useful to me as part of the drag-and-drop support for TabItem, but for the sake of completeness (and also for anyone who may be inheriting the TabItem class), well, here you go!

Also marked most items in TabDisplayItem not receiving XML comments, and rearranged the drag-and-drop code a bit. Also also added missing comments to TabItem.