Limit SelectPanel move functions to not work with ItemsSource
While ItemsSource is not set to the control's internal items list, the move functions won't do anything anyway (as they're tied to using the internal items list). Thus, it's better to simply exit quickly. Perhaps in the future, I can modify it to work with certain IEnumerable types in the ItemsSource, we'll see.

Also added remaining missing comments, including comments for the convenience methods previously marked as obsolete.