Add error handling to DoubleStringConverter
Not sure why this class is still here, as I don't think I use it anywhere, but whatever lol! I wonder how it ended up here in the first place...

Anyway, added handling in case the parameter passed in was not an integer (or couldn't be casted to an integer).