Create ListEditorDialog, refactor PropertyList
This is a big commit! I've started the practice of trying to break my commits into smaller parts, but a lot of this tied into each other, sooooo yeah.

I updated the code that PropertyList uses to try to get an editor for a property; now, it'll attempt to read from the list of registered editors for a direct match first, before attempting some other things. I've also exposed this function publicly, so that others (including other classes of mine) can use PropertyList's logic to get a valid property editor, if possible.

I've added a ParentPropertyList property to IPropertyEditor, so that some editors have access to the PropertyList to use that specific function. I've added this property to all IPropertyEditor implementers so far.

Now, we come to the big piece here: I've created a ListEditorDialog, which will display the elements present in an IEnumerable or List. If it is a List, it'll allow adding, editing, and removing items as well. If it isn't a basic type, then you can view the type's properties in the embedded PropertyList on the right (the nesting, it begins!). This is a powerful window and feature to add onto the PropertyList, and I'm excited to have it here now!

Note that right now, it will only support adding items to a list, if that item has a parameterless constructor... which, apparently, strings do not. So List<string> cannot be added to right now!

There's probably some cleaning up of my code that I have to do before it's all ready for release. A lot of this was put together as I went along, so yeah.