Add special handling for List<string> in ListEditorDialog
When loading in a List<string>, then the ListEditorDialog goes into String List mode, which overrides the handling for adding in an item, and allows new strings to be added.

Also adds some extra error handling if the list modifies its contents while the dialog is open. This does not update the appearance of the list (I may add a Refresh button), but it at least prevents a crash.

Also added in a Grid with subelements for testing in the PropertyListTest.