Fixed issues when using ItemsSource with SelectPanel
Fixed visual and other issues with the SelectPanel's items looking and acting weird when the ItemsSource is set to an external item. This was primarily because a lot of the internal code was still connected to using the internal Items collection.

Note that if the ItemsSource is not set to a SelectableCollection maintained elsewhere, then there may be some functionality not working as expected. I'll do some quick tests on that next.