Implement ICollection<T> in SelectableCollectionView
Not something I really wanted to do, but for the ability to add it as an ItemsSource in SelectPanel, that's what I've had to do!

It is set up similarly to a ReadOnlyCollection, where attempting to add or insert items causes a NotSupportedException. Removing, however, does not (because ListCollectionView has a Remove method for some reason).

Also added a new constructor that can accept a non-generic IList. However, the actual passed in list must still implement IList<T>.